+++ front/lib/actions/mcp_internal_actions/servers/interactive_content/index.ts
@@ -12,12 +12,10 @@ import {
   RETRIEVE_INTERACTIVE_CONTENT_FILE_TOOL_NAME,
   REVERT_INTERACTIVE_CONTENT_FILE_TOOL_NAME,
 } from "@app/lib/actions/mcp_internal_actions/servers/interactive_content/types";
-import { makeCoreSearchNodesFilters } from "@app/lib/actions/mcp_internal_actions/tools/utils";
+import { fetchTemplateContent } from "@app/lib/actions/mcp_internal_actions/servers/interactive_content/template_utils";
 import { makeInternalMCPServer } from "@app/lib/actions/mcp_internal_actions/utils";
 import { withToolLogging } from "@app/lib/actions/mcp_internal_actions/wrappers";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
-import { getSkillDataSourceConfigurations } from "@app/lib/api/assistant/skill_actions";
-import config from "@app/lib/api/config";
 import {
   createClientExecutableFile,
   editClientExecutableFile,
@@ -28,18 +26,9 @@ import {
 } from "@app/lib/api/files/client_executable";
 import { formatValidationWarningsForLLM } from "@app/lib/api/files/content_validation";
 import type { Authenticator } from "@app/lib/auth";
-import { DataSourceViewResource } from "@app/lib/resources/data_source_view_resource";
 import type { FileResource } from "@app/lib/resources/file_resource";
-import { SkillResource } from "@app/lib/resources/skill/skill_resource";
-import logger from "@app/logger/logger";
 import type { InteractiveContentFileContentType } from "@app/types";
-import {
-  CoreAPI,
-  Err,
-  frameContentType,
-  INTERACTIVE_CONTENT_FILE_FORMATS,
-  Ok,
-} from "@app/types";
+import { Err, frameContentType, INTERACTIVE_CONTENT_FILE_FORMATS, Ok } from "@app/types";
 
 const MAX_FILE_SIZE_BYTES = 1 * 1024 * 1024; // 1MB
 
@@ -158,8 +147,6 @@ function createServer(
 
         // Fetch template content if mode is 'template'
         if (mode === "template") {
-          const templateNodeId = source;
-
           if (!agentLoopContext?.runContext) {
             return new Err(
               new MCPError(
@@ -169,139 +156,17 @@ function createServer(
             );
           }
 
-          const { agentConfiguration, conversation } =
-            agentLoopContext.runContext;
+          const templateResult = await fetchTemplateContent(
+            auth,
+            agentLoopContext.runContext,
+            { templateNodeId: source }
+          );
 
-          // Fetch skills for this agent and conversation (same pattern as agent loop).
-          const { enabledSkills } = await SkillResource.listForAgentLoop(auth, {
-            agentConfiguration,
-            conversation,
-          });
-
-          // Get merged data source configurations from skills.
-          const dataSourceConfigurations =
-            await getSkillDataSourceConfigurations(auth, {
-              skills: enabledSkills,
-            });
-
-          if (dataSourceConfigurations.length === 0) {
-            return new Err(
-              new MCPError(
-                "No data sources found in skills configuration. " +
-                  "Template nodes can only be used when the agent has skills with attached knowledge.",
-                { tracked: false }
-              )
-            );
+          if (templateResult.isErr()) {
+            return templateResult;
           }
 
-          // Resolve DataSourceConfiguration[] to ResolvedDataSourceConfiguration[].
-          const agentDataSourceConfigurations = [];
-          for (const config of dataSourceConfigurations) {
-            const dataSourceView = await DataSourceViewResource.fetchById(
-              auth,
-              config.dataSourceViewId
-            );
-
-            if (!dataSourceView) {
-              return new Err(
-                new MCPError(
-                  `Data source view not found: ${config.dataSourceViewId}`,
-                  { tracked: false }
-                )
-              );
-            }
-
-            const dataSource = dataSourceView.dataSource;
-
-            agentDataSourceConfigurations.push({
-              ...config,
-              dataSource: {
-                dustAPIProjectId: dataSource.dustAPIProjectId,
-                dustAPIDataSourceId: dataSource.dustAPIDataSourceId,
-                connectorProvider: dataSource.connectorProvider,
-                name: dataSource.name,
-              },
-              dataSourceView,
-            });
-          }
-
-          // Search for the template node
-          const coreAPI = new CoreAPI(config.getCoreAPIConfig(), logger);
-          const searchResult = await coreAPI.searchNodes({
-            filter: {
-              node_ids: [templateNodeId],
-              data_source_views: makeCoreSearchNodesFilters({
-                agentDataSourceConfigurations,
-              }),
-            },
-          });
-
-          if (searchResult.isErr() || searchResult.value.nodes.length === 0) {
-            return new Err(
-              new MCPError(
-                `Could not find template node: ${templateNodeId}. ${
-                  searchResult.isErr()
-                    ? `Error: ${searchResult.error.message}`
-                    : "The node may not exist or may not be accessible through your skills configuration."
-                }`,
-                { tracked: false }
-              )
-            );
-          }
-
-          const node = searchResult.value.nodes[0];
-
-          if (node.node_type !== "document") {
-            return new Err(
-              new MCPError(
-                `Template node is of type ${node.node_type}, not a document. Only document nodes can be used as templates.`,
-                { tracked: false }
-              )
-            );
-          }
-
-          // Get dataSource from the data source configuration
-          const dataSource = agentDataSourceConfigurations.find(
-            (config) =>
-              config.dataSource.dustAPIDataSourceId === node.data_source_id
-          )?.dataSource;
-
-          if (!dataSource) {
-            return new Err(
-              new MCPError(
-                `Could not find data source for template node: ${templateNodeId}`,
-                { tracked: false }
-              )
-            );
-          }
-
-          // Read the template node content
-          const readResult = await coreAPI.getDataSourceDocumentText({
-            dataSourceId: node.data_source_id,
-            documentId: node.node_id,
-            projectId: dataSource.dustAPIProjectId,
-          });
-
-          if (readResult.isErr()) {
-            return new Err(
-              new MCPError(
-                `Could not read template node: ${templateNodeId}. Error: ${readResult.error.message}`,
-                { tracked: false }
-              )
-            );
-          }
-
-          fileContent = readResult.value.text;
-
-          // Validate template content size
-          if (fileContent.length > MAX_FILE_SIZE_BYTES) {
-            return new Err(
-              new MCPError(
-                `Template content is too large (${fileContent.length} bytes). Maximum size is ${MAX_FILE_SIZE_BYTES} bytes (1MB).`,
-                { tracked: false }
-              )
-            );
-          }
+          fileContent = templateResult.value;
         } else {
           // Inline content path (mode === 'inline')
           fileContent = source;
