+++ front/lib/actions/mcp_internal_actions/servers/interactive_content/template_utils.ts
@@ -1,0 +1,154 @@
+import type { AgentLoopRunContextType } from "@app/lib/actions/types";
+import { getSkillDataSourceConfigurations } from "@app/lib/api/assistant/skill_actions";
+import config from "@app/lib/api/config";
+import type { Authenticator } from "@app/lib/auth";
+import { DataSourceViewResource } from "@app/lib/resources/data_source_view_resource";
+import { SkillResource } from "@app/lib/resources/skill/skill_resource";
+import logger from "@app/logger/logger";
+import type { Result } from "@app/types";
+import { CoreAPI, Err, Ok } from "@app/types";
+
+import { MCPError } from "../../../mcp_errors";
+import { makeCoreSearchNodesFilters } from "../../tools/utils";
+
+const MAX_TEMPLATE_SIZE_BYTES = 1 * 1024 * 1024; // 1MB
+
+/**
+ * Fetches template content from a knowledge node using the agent's skills configuration.
+ */
+export async function fetchTemplateContent(
+  auth: Authenticator,
+  runContext: AgentLoopRunContextType,
+  { templateNodeId }: { templateNodeId: string }
+): Promise<Result<string, MCPError>> {
+  const { agentConfiguration, conversation } = runContext;
+
+  // Fetch skills for this agent and conversation (same pattern as agent loop).
+  const { enabledSkills } = await SkillResource.listForAgentLoop(auth, {
+    agentConfiguration,
+    conversation,
+  });
+
+  // Get merged data source configurations from skills.
+  const dataSourceConfigurations = await getSkillDataSourceConfigurations(
+    auth,
+    {
+      skills: enabledSkills,
+    }
+  );
+
+  if (dataSourceConfigurations.length === 0) {
+    return new Err(
+      new MCPError(
+        "No data sources found in skills configuration. " +
+          "Template nodes can only be used when the agent has skills with attached knowledge.",
+        { tracked: false }
+      )
+    );
+  }
+
+  // Resolve DataSourceConfiguration[] to ResolvedDataSourceConfiguration[].
+  const agentDataSourceConfigurations = [];
+  for (const config of dataSourceConfigurations) {
+    const dataSourceView = await DataSourceViewResource.fetchById(
+      auth,
+      config.dataSourceViewId
+    );
+
+    if (!dataSourceView) {
+      return new Err(
+        new MCPError(`Data source view not found: ${config.dataSourceViewId}`, {
+          tracked: false,
+        })
+      );
+    }
+
+    const dataSource = dataSourceView.dataSource;
+
+    agentDataSourceConfigurations.push({
+      ...config,
+      dataSource: {
+        dustAPIProjectId: dataSource.dustAPIProjectId,
+        dustAPIDataSourceId: dataSource.dustAPIDataSourceId,
+        connectorProvider: dataSource.connectorProvider,
+        name: dataSource.name,
+      },
+      dataSourceView,
+    });
+  }
+
+  // Search for the template node.
+  const coreAPI = new CoreAPI(config.getCoreAPIConfig(), logger);
+  const searchResult = await coreAPI.searchNodes({
+    filter: {
+      node_ids: [templateNodeId],
+      data_source_views: makeCoreSearchNodesFilters({
+        agentDataSourceConfigurations,
+      }),
+    },
+  });
+
+  if (searchResult.isErr() || searchResult.value.nodes.length === 0) {
+    return new Err(
+      new MCPError(
+        `Could not find template node: ${templateNodeId}. ${
+          searchResult.isErr()
+            ? `Error: ${searchResult.error.message}`
+            : "The node may not exist or may not be accessible through your skills configuration."
+        }`,
+        { tracked: false }
+      )
+    );
+  }
+
+  const node = searchResult.value.nodes[0];
+  if (node.node_type !== "document") {
+    return new Err(
+      new MCPError(
+        `Template node is of type ${node.node_type}, not a document. Only document nodes can be used as templates.`,
+        { tracked: false }
+      )
+    );
+  }
+
+  // Get dataSource from the data source configuration.
+  const dataSource = agentDataSourceConfigurations.find(
+    (config) => config.dataSource.dustAPIDataSourceId === node.data_source_id
+  )?.dataSource;
+
+  if (!dataSource) {
+    return new Err(
+      new MCPError(
+        `Could not find data source for template node: ${templateNodeId}`,
+        { tracked: false }
+      )
+    );
+  }
+
+  // Read the template node content.
+  const readResult = await coreAPI.getDataSourceDocumentText({
+    dataSourceId: node.data_source_id,
+    documentId: node.node_id,
+    projectId: dataSource.dustAPIProjectId,
+  });
+  if (readResult.isErr()) {
+    return new Err(
+      new MCPError(
+        `Could not read template node: ${templateNodeId}. Error: ${readResult.error.message}`,
+        { tracked: false }
+      )
+    );
+  }
+
+  const templateContent = readResult.value.text;
+  if (templateContent.length > MAX_TEMPLATE_SIZE_BYTES) {
+    return new Err(
+      new MCPError(
+        `Template content is too large (${templateContent.length} bytes). Maximum size is ${MAX_TEMPLATE_SIZE_BYTES} bytes (1MB).`,
+        { tracked: false }
+      )
+    );
+  }
+
+  return new Ok(templateContent);
+}
